Update package script

+++ bin/npm_publish.sh
@@ -1,5 +1,12 @@
 #!/usr/bin/env sh
 
+git diff-index --quiet HEAD --
+if [ $? -ne 0 ]
+then
+  echo "Uncommitted git changes"
+  exit 1
+fi
+
 rm -rf build/publish
 mkdir -p build/publish/bin
 
@@ -12,4 +19,13 @@ cp README.md build/publish/README.md
 
 cd build/publish
 npm publish .
+
+PACKAGE_VERSION=$(cat package.json \
+  | grep version \
+  | head -1 \
+  | awk -F: '{ print $2 }' \
+  | sed 's/[",]//g' \
+  | tr -d '[[:space:]]')
+
+git tag $PACKAGE_VERSION && git push --tags
